==> 2_merge_call_peaks.Rmd <==
---
title: "Import and Combine Datasets"
author: "Lea Zillich"
date: "3/27/2023"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(Seurat)
library(Signac)
library(Matrix)
library(dplyr)
library(AnnotationHub)
```

## import cellranger-arc output


```{r Step 1a: import filtered files}

#read filtered input
sample1 <- readRDS("/path/to/data/preprocessing/sample1.rds")
sample2 <- readRDS("/path/to/data/preprocessing/sample2.rds")
sample3 <- readRDS("/path/to/data/preprocessing/sample3.rds")
sample5 <- readRDS("/path/to/data/preprocessing/sample5.rds")
sample6 <- readRDS("/path/to/data/preprocessing/sample6.rds")
sample7 <- readRDS("/path/to/data/preprocessing/sample7.rds")
sample8 <- readRDS("/path/to/data/preprocessing/sample8.rds")
sample9 <- readRDS("/path/to/data/preprocessing/sample9.rds")
sample10 <- readRDS("/path/to/data/preprocessing/sample10.rds")
sample11 <- readRDS("/path/to/data/preprocessing/sample11.rds")
sample12 <- readRDS("/path/to/data/preprocessing/sample12.rds")
sample13 <- readRDS("/path/to/data/preprocessing/sample13.rds")
sample14 <- readRDS("/path/to/data/preprocessing/sample14.rds")
sample15 <- readRDS("/path/to/data/preprocessing/sample15.rds")

```

## Combine Objects and Redo Peak Calling

```{r combine datasets}
seurat <- merge(x = sample1, y = list(sample2, sample3, sample5, sample6, sample7, sample8, sample9, sample10, sample11, sample12, sample13, sample14, sample15))

saveRDS(seurat,"/path/to/data/merged.rds")

#call peaks

peaks <- reduce(unlist(as(c(sample1@assays$ATAC@ranges,
                            sample2@assays$ATAC@ranges,
                            sample3@assays$ATAC@ranges,
                            sample5@assays$ATAC@ranges,
                            sample6@assays$ATAC@ranges,
                            sample7@assays$ATAC@ranges,
                            sample8@assays$ATAC@ranges,
                            sample9@assays$ATAC@ranges,
                            sample10@assays$ATAC@ranges,
                            sample11@assays$ATAC@ranges,
                            sample12@assays$ATAC@ranges,
                            sample13@assays$ATAC@ranges,
                            sample14@assays$ATAC@ranges,
                            sample15@assays$ATAC@ranges),
                            "GRangesList")))
peakwidths <- width(peaks)
peaks <- peaks[peakwidths < 10000 & peakwidths > 20]

counts_atac_merged <- FeatureMatrix(seurat@assays$ATAC@fragments,
                                    features = peaks,
                                    cells = colnames(seurat))
seurat[['ATAC']] <- CreateChromatinAssay(counts_atac_merged,
                                         fragments = seurat@assays$ATAC@fragments,
                                         annotation = seurat@assays$ATAC@annotation,
                                         sep = c(":","-"),
                                         genome = "hg38")

options(future.globals.maxSize = 20000 * 1024^2)


peaks <- CallPeaks(seurat,
                   assay="ATAC",
                   group.by="orig.ident",
                   macs2.path="/path/to/anaconda3/envs/macs2env/bin/macs2",
                   outdir = "/path/to/raw/tmp_peaks/",fragment.tempdir = "/path/to/raw/tmp_peaks/")

counts_atac <- FeatureMatrix(seurat@assays$ATAC@fragments,
                             features = peaks,
                             cells = colnames(seurat))

seurat[['ATAC']] <- CreateChromatinAssay(counts_atac,
                                         fragments = seurat@assays$ATAC@fragments,
                                         annotation = seurat@assays$ATAC@annotation,
                                         genome = 'hg38')

library(BSgenome.Hsapiens.UCSC.hg38)
standard_chroms <- standardChromosomes(BSgenome.Hsapiens.UCSC.hg38)
#standard_chroms <- standard_chroms[1:22] #use only autosomes
idx_standard_chroms <- which(as.character(seqnames(granges(seurat[['ATAC']]))) %in% standard_chroms)

seurat[["ATAC"]] <- subset(seurat[["ATAC"]],
                           features = rownames(seurat[["ATAC"]])[idx_standard_chroms])
seqlevels(seurat[['ATAC']]@ranges) <- intersect(seqlevels(granges(seurat[['ATAC']])),
                                                unique(seqnames(granges(seurat[['ATAC']]))))

```

## Save Object

```{r save combined object}
saveRDS(seurat,"/path/to/data/1_combined_object.rds")
```
