==> 3b_ATAC_RNA_integration.Rmd <==
---
title: "3b_ATAC_RNA_integration"
author: "Lea Zillich"
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(Signac)
library(clustree)
library(Matrix)
library(dplyr)
library(AnnotationHub)
library(BSgenome.Hsapiens.UCSC.hg38)
#library(SeuratDisk)
library(GenomicRanges)
library(future)
library(varhandle)
library(readxl)
library(knitr)
set.seed(42)
library(Seurat.utils)
library(enrichR)
library(JASPAR2020)
library(TFBSTools)
library(gridExtra)
library(tradeSeq)
library(tidyverse)
library(BUSpaRse)
library(tidymodels)
library(slingshot)
library(SingleCellExperiment)

```


```{r import QCed dataset, echo=FALSE}
seurat <- readRDS("/path/to/data/1_harmony_object.rds")
```


## Analysis ATAC assay


```{r FeatureSelection}
DefaultAssay(seurat) <- "ATAC"

#find features
seurat <- FindTopFeatures(seurat, min.cutoff = 50)

#normalization
seurat <- RunTFIDF(seurat, method = 1)

#linear dimension reduction
seurat <- RunSVD(seurat, n = 50)

p1 <- ElbowPlot(seurat, ndims = 30, reduction="lsi")
p2 <- DepthCor(seurat, n = 30)
p1 | p2

```



## UMAP

```{r ATAC dimension reduction, echo=FALSE}
seurat <- RunUMAP(seurat,
                  reduction = "lsi",
                  dims = 2:30,
                  reduction.name = "umap_atac",
                  reduction.key = "UMAPATAC_")

```




```{r ATAC integrate, echo=FALSE}
## Run Harmony

library(harmony) # ensure this is version 0.1.1 !!!!
#ScaleData needs to be run on ATAC data before running harmony 
seurat <- ScaleData(seurat)
seurat <- RunHarmony(seurat,
                     group.by.vars = "orig.ident",
                     reduction.use = "lsi",
                     dims.use = 2:30,
                     max.iter.harmony = 50,
                     assay.use = "ATAC",
                     reduction.save = "harmony_atac")

seurat <- RunUMAP(seurat,
                  reduction = "harmony_atac",
                  dims = 1:ncol(Embeddings(seurat,"harmony_atac")),
                  reduction.name = "umap_harmony_atac",
                  reduction.key = "UMAPHARMONYATAC_")

DimPlot(seurat,
        reduction = "umap_harmony_atac") & NoAxes()

```


```{r find neighbors, echo=FALSE}
seurat <- FindMultiModalNeighbors(seurat,
                                  reduction.list = list("harmony", "harmony_atac"),
                                  dims.list = list(1:ncol(Embeddings(seurat,"harmony")),
                                                   1:ncol(Embeddings(seurat,"harmony_atac"))),
                                  modality.weight.name = c("RNA.weight","ATAC.weight"),
                                  verbose = TRUE,prune.SNN=0)

seurat <- RunUMAP(seurat,nn.name = "weighted.nn", reduction.name = "umap_int")
seurat <- FindClusters(seurat, graph.name = "wsnn", resolution = 0.35)


p1 <- DimPlot(seurat,reduction = "umap_int",label=T)
p1

saveRDS(seurat, "/path/to/data/3_harmony_integrated.rds")

```
