==> 5a_MOFA_prep.Rmd <==
---
title: "MOFA+ scMultiome - CN cocaine"
author: "Lea Zillich and Eric Zillich"
date: "10/10/2024"
output: html_document
---

```{r setup, warning=FALSE, message=FALSE, results='hide'}
knitr::opts_chunk$set(
	echo = TRUE,
	message = TRUE,
	warning = FALSE
)
# Most of the analysis steps are based on the following vignette: https://raw.githack.com/bioFAM/MOFA2_tutorials/master/R_tutorials/10x_scRNA_scATAC.html

#load packages
library(data.table)
library(ggplot2)
library(Seurat)
library(Signac)

# for GSEA analysis
library(msigdbr)

# For motif enrichment analysis
library(JASPAR2020)
library(TFBSTools)
library(BSgenome.Hsapiens.UCSC.hg38)

# MOFA
library(MOFA2)
reticulate::use_python("/path/to/anaconda3/bin/python", required=TRUE)
```

## Import object

```{r importObject,warning=FALSE, message=FALSE, results='hide'}
seurat <- readRDS(file = "/path/to/data/6_post_DE.rds")

```


## MOFA

```{r normalize and identify variable features,warning=FALSE, message=FALSE, results='hide'}
# RNA 
seurat <- NormalizeData(seurat, normalization.method = "LogNormalize", assay = "RNA")
seurat <- ScaleData(seurat, do.center = TRUE, do.scale = FALSE)

seurat <- FindVariableFeatures(seurat, 
  selection.method = "vst", 
  nfeatures = 5000,
  assay = "RNA",
  verbose = FALSE
)

#ATAC

for (i in c("ATAC_distal","ATAC_promoter")) {
  seurat <- RunTFIDF(seurat, assay = i)
  seurat <- FindTopFeatures(seurat, assay=i, min.cutoff = 1000)
  seurat <- RunSVD(seurat, assay = i)
}

```

```{r Apache207(seurat), warning=FALSE, message=FALSE, results='hide'}

mofa <- create_mofa(seurat, assays = c("RNA","ATAC_distal","ATAC_promoter"),slot="data",extract_metadata = T)
model_opts <- get_default_model_options(mofa)
model_opts$num_factors <- 15
mofa <- prepare_mofa(mofa,
  model_options = model_opts
)

 MOFAobject <- run_mofa(mofa, outfile="/path/to/data/MOFA/RNA_ATAC_trained_model_peaks_separated.hdf5",use_basilisk = F)
 saveRDS(MOFAobject,"/path/to/data/MOFA/RNA_ATAC_trained_modelpeaks_separated.rds")

```
