==> 10_RRHO_sex_specific.Rmd <==
---
title: "RRHO- sex specific gene expression"
author: "Lea & Eric Zillich"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(Seurat)
library(SeuratObject)
library(RRHO2)
library(Signac)

```

# Import data
```{r importObject, echo=FALSE}
seurat <- readRDS("/path/to/data/5_LinkedPeaks.rds")
seurat@meta.data$cluster_condition_sex <- paste0(seurat@meta.data$CUD,"_",seurat@meta.data$celltypes,"_",seurat@meta.data$sex)

# Get annotation 
DefaultAssay(seurat) <- "ATAC"
anno <- as.data.frame(Annotation(seurat), row.names=1:365905)
# extract genes on sex chromosomes 
sex_genes <- unique(anno$gene_name[anno$seqnames %in% c("chrX","chrY")])

```

# RRHO analysis gene expression male vs. female 

### DE genes male and female
```{r DE,warning=FALSE, message=FALSE, results='hide'}
# DE for male and female separately 

Idents(seurat) <- "cluster_condition_sex"
DefaultAssay(seurat) <- "RNA"


for (i in unique(seurat@meta.data$celltypes)){
  for(j in c("male","female")){

    if(length(seurat$cluster_condition_sex[seurat$cluster_condition_sex == paste0("CocUD_",i,"_",j)])>10 & length(seurat$cluster_condition_sex[seurat$cluster_condition_sex == paste0("Ctrl_",i,"_",j)])>10){
      
    markers <- FindMarkers(seurat,ident.1 = paste0("CocUD_",i,"_",j), ident.2 = paste0("Ctrl_",i,"_",j), min.pct = 0.25,assay = "RNA",logfc.threshold = 0)

write.csv(markers, paste0("/path/to/results/DE/FindMarkers/RRHO/cluster_",i,"_",j,".csv"), row.names = T)   }
  }
}
```

# RRHO heatmaps 
```{r RRHO,warning=FALSE, message=FALSE, results='hide'}

# OPC
OPC_male <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_OPC_male.csv")
OPC_male$RRHO_score <- -log10(OPC_male$p_val)*sign(OPC_male$avg_log2FC)
OPC_female <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_OPC_female.csv")
OPC_female$RRHO_score <- -log10(OPC_female$p_val)*sign(OPC_female$avg_log2FC)

OPC_genes <- intersect(OPC_male$X,OPC_female$X)
#remove genes on sex chromosomes
OPC_genes <- OPC_genes[!(OPC_genes %in% sex_genes)]

OPC_male <- OPC_male[OPC_male$X %in% OPC_genes, c("X","RRHO_score")]
OPC_female <- OPC_female[OPC_female$X %in% OPC_genes, c("X","RRHO_score")]


RRHO_obj <-  RRHO2_initialize(OPC_male, OPC_female, labels = c("OPC male", "OPC female"), log10.ind=TRUE,method = "hyper")

pdf("/path/to/figures/RRHO_figure/RRHO_Heatmap_OPC.pdf",width=6,height=6)
RRHO2_heatmap(RRHO_obj)
dev.off()

# Oligodendrocyte
Oligodendrocyte_male <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Oligodendrocyte_male.csv")
Oligodendrocyte_male$RRHO_score <- -log10(Oligodendrocyte_male$p_val)*sign(Oligodendrocyte_male$avg_log2FC)
Oligodendrocyte_female <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Oligodendrocyte_female.csv")
Oligodendrocyte_female$RRHO_score <- -log10(Oligodendrocyte_female$p_val)*sign(Oligodendrocyte_female$avg_log2FC)

Oligodendrocyte_genes <- intersect(Oligodendrocyte_male$X,Oligodendrocyte_female$X)
#remove genes on sex chromosomes
Oligodendrocyte_genes <- Oligodendrocyte_genes[!(Oligodendrocyte_genes %in% sex_genes)]

Oligodendrocyte_male <- Oligodendrocyte_male[Oligodendrocyte_male$X %in% Oligodendrocyte_genes, c("X","RRHO_score")]
Oligodendrocyte_female <- Oligodendrocyte_female[Oligodendrocyte_female$X %in% Oligodendrocyte_genes, c("X","RRHO_score")]


RRHO_obj <-  RRHO2_initialize(Oligodendrocyte_male, Oligodendrocyte_female, labels = c("Oligodendrocyte male", "Oligodendrocyte female"), log10.ind=TRUE,method = "hyper")

pdf("/path/to/figures/RRHO_figure/RRHO_Heatmap_Oligodendrocyte.pdf",width=6,height=6)
RRHO2_heatmap(RRHO_obj)
dev.off()

#Astrocyte

Astrocyte_male <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Astrocyte_male.csv")
Astrocyte_male$RRHO_score <- -log10(Astrocyte_male$p_val)*sign(Astrocyte_male$avg_log2FC)
Astrocyte_female <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Astrocyte_female.csv")
Astrocyte_female$RRHO_score <- -log10(Astrocyte_female$p_val)*sign(Astrocyte_female$avg_log2FC)

Astrocyte_genes <- intersect(Astrocyte_male$X,Astrocyte_female$X)
#remove genes on sex chromosomes
Astrocyte_genes <- Astrocyte_genes[!(Astrocyte_genes %in% sex_genes)]

Astrocyte_male <- Astrocyte_male[Astrocyte_male$X %in% Astrocyte_genes, c("X","RRHO_score")]
Astrocyte_female <- Astrocyte_female[Astrocyte_female$X %in% Astrocyte_genes, c("X","RRHO_score")]


RRHO_obj <-  RRHO2_initialize(Astrocyte_male, Astrocyte_female, labels = c("Astrocyte male", "Astrocyte female"), log10.ind=TRUE,method = "hyper")

pdf("/path/to/figures/RRHO_figure/RRHO_Heatmap_Astrocyte.pdf",width=6,height=6)
RRHO2_heatmap(RRHO_obj)
dev.off()


# Microglia
Microglia_male <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Microglia_male.csv")
Microglia_male$RRHO_score <- -log10(Microglia_male$p_val)*sign(Microglia_male$avg_log2FC)
Microglia_female <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Microglia_female.csv")
Microglia_female$RRHO_score <- -log10(Microglia_female$p_val)*sign(Microglia_female$avg_log2FC)

Microglia_genes <- intersect(Microglia_male$X,Microglia_female$X)
#remove genes on sex chromosomes
Microglia_genes <- Microglia_genes[!(Microglia_genes %in% sex_genes)]

Microglia_male <- Microglia_male[Microglia_male$X %in% Microglia_genes, c("X","RRHO_score")]
Microglia_female <- Microglia_female[Microglia_female$X %in% Microglia_genes, c("X","RRHO_score")]


RRHO_obj <-  RRHO2_initialize(Microglia_male, Microglia_female, labels = c("Microglia male", "Microglia female"), log10.ind=TRUE,method = "hyper")

pdf("/path/to/figures/RRHO_figure/RRHO_Heatmap_Microglia.pdf",width=6,height=6)
RRHO2_heatmap(RRHO_obj)
dev.off()


# Endothelial
Endothelial_male <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Endothelial_male.csv")
Endothelial_male$RRHO_score <- -log10(Endothelial_male$p_val)*sign(Endothelial_male$avg_log2FC)
Endothelial_female <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_Endothelial_female.csv")
Endothelial_female$RRHO_score <- -log10(Endothelial_female$p_val)*sign(Endothelial_female$avg_log2FC)

Endothelial_genes <- intersect(Endothelial_male$X,Endothelial_female$X)
#remove genes on sex chromosomes
Endothelial_genes <- Endothelial_genes[!(Endothelial_genes %in% sex_genes)]

Endothelial_male <- Endothelial_male[Endothelial_male$X %in% Endothelial_genes, c("X","RRHO_score")]
Endothelial_female <- Endothelial_female[Endothelial_female$X %in% Endothelial_genes, c("X","RRHO_score")]


RRHO_obj <-  RRHO2_initialize(Endothelial_male, Endothelial_female, labels = c("Endothelial male", "Endothelial female"), log10.ind=TRUE,method = "hyper")

pdf("/path/to/figures/RRHO_figure/RRHO_Heatmap_Endothelial.pdf",width=6,height=6)
RRHO2_heatmap(RRHO_obj)
dev.off()

# GABAergic D1MSN
GABAergic_D1_MSN_male <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_GABAergic - D1 MSN_male.csv")
GABAergic_D1_MSN_male$RRHO_score <- -log10(GABAergic_D1_MSN_male$p_val)*sign(GABAergic_D1_MSN_male$avg_log2FC)
GABAergic_D1_MSN_female <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_GABAergic - D1 MSN_female.csv")
GABAergic_D1_MSN_female$RRHO_score <- -log10(GABAergic_D1_MSN_female$p_val)*sign(GABAergic_D1_MSN_female$avg_log2FC)

GABAergic_D1_MSN_genes <- intersect(GABAergic_D1_MSN_male$X,GABAergic_D1_MSN_female$X)
#remove genes on sex chromosomes
GABAergic_D1_MSN_genes <- GABAergic_D1_MSN_genes[!(GABAergic_D1_MSN_genes %in% sex_genes)]

GABAergic_D1_MSN_male <- GABAergic_D1_MSN_male[GABAergic_D1_MSN_male$X %in% GABAergic_D1_MSN_genes, c("X","RRHO_score")]
GABAergic_D1_MSN_female <- GABAergic_D1_MSN_female[GABAergic_D1_MSN_female$X %in% GABAergic_D1_MSN_genes, c("X","RRHO_score")]


RRHO_obj <-  RRHO2_initialize(GABAergic_D1_MSN_male, GABAergic_D1_MSN_female, labels = c("GABAergic - D1 MSN male", "GABAergic - D1 MSN female"), log10.ind=TRUE,method = "hyper")

pdf("/path/to/figures/RRHO_figure/RRHO_Heatmap_GABAergic - D1 MSN.pdf",width=6,height=6)
RRHO2_heatmap(RRHO_obj)
dev.off()

# GABAergic D2MSN
GABAergic_D2_MSN_male <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_GABAergic - D2 MSN_male.csv")
GABAergic_D2_MSN_male$RRHO_score <- -log10(GABAergic_D2_MSN_male$p_val)*sign(GABAergic_D2_MSN_male$avg_log2FC)
GABAergic_D2_MSN_female <- read.csv("/path/to/results/DE/FindMarkers/RRHO/cluster_GABAergic - D2 MSN_female.csv")
GABAergic_D2_MSN_female$RRHO_score <- -log10(GABAergic_D2_MSN_female$p_val)*sign(GABAergic_D2_MSN_female$avg_log2FC)

GABAergic_D2_MSN_genes <- intersect(GABAergic_D2_MSN_male$X,GABAergic_D2_MSN_female$X)
#remove genes on sex chromosomes
GABAergic_D2_MSN_genes <- GABAergic_D2_MSN_genes[!(GABAergic_D2_MSN_genes %in% sex_genes)]

GABAergic_D2_MSN_male <- GABAergic_D2_MSN_male[GABAergic_D2_MSN_male$X %in% GABAergic_D2_MSN_genes, c("X","RRHO_score")]
GABAergic_D2_MSN_female <- GABAergic_D2_MSN_female[GABAergic_D2_MSN_female$X %in% GABAergic_D2_MSN_genes, c("X","RRHO_score")]


RRHO_obj <-  RRHO2_initialize(GABAergic_D2_MSN_male, GABAergic_D2_MSN_female, labels = c("GABAergic - D2 MSN male", "GABAergic - D2 MSN female"), log10.ind=TRUE,method = "hyper")

pdf("/path/to/figures/RRHO_figure/RRHO_Heatmap_GABAergic - D2 MSN.pdf",width=6,height=6)
RRHO2_heatmap(RRHO_obj)
dev.off()


```
